These are the changes I needed to make whilst testing SDO for C++ with Guththila. There are a few signed/unsigned problems, one return of an integer where an object is expected, and the problem of reallocation of the read buffer.

+++ modules/xml/parser/guththila/impl/src/guththila_string.c
@@ -83,7 +83,7 @@ guththila_strndup (guththila_environment_t *environment,
         register guththila_char_t *dst;
 
         n = guththila_strlen (s1);
-        if (len < n)
+        if ((size_t)len < n)
             n = len;
         dst = (guththila_char_t *) GUTHTHILA_MALLOC(environment->allocator ,(n + 1));
         if (dst)
